@@ -1,6 +1,6 @@
 {
   "name": "eximiabots-radiox",
-  "version": "0.5.7",
+  "version": "0.5.8",
   "description": "Internet Radio to your Discord guild",
   "main": "index.js",
   "scripts": {
@@ -19,14 +19,14 @@
   },
   "dependencies": {
     "@discordjs/voice": "^0.17.0",
-    "discord.js": "^14.15.3",
+    "discord.js": "^14.16.3",
     "dotenv": "^16.4.5",
-    "libsodium-wrappers": "^0.7.13",
+    "libsodium-wrappers": "^0.7.15",
     "path": "^0.12.7"
   },
   "devDependencies": {
-    "rimraf": "^5.0.7",
-    "typescript": "^5.4.5"
+    "rimraf": "^6.0.1",
+    "typescript": "^5.6.2"
   },
   "engines": {
       "node": ">=20.0.0",
