@@ -3,16 +3,15 @@ import RadioClient from "../../Client";
 
 export default function interactionCreate(client: RadioClient, interaction: Interaction) {
     if(!(interaction.isButton()) && !(interaction.isChatInputCommand()) && !(interaction.isStringSelectMenu())) return;
+    if(interaction.channel?.type == ChannelType.DM || interaction.channel?.type == ChannelType.GroupDM) return;
 
-    if(interaction.channel?.type != ChannelType.DM){
-        const permissions = interaction.channel?.permissionsFor(interaction.client.user);
-        if (!permissions?.has(PermissionFlagsBits.ViewChannel)) return;
+    const permissions = interaction.channel?.permissionsFor(interaction.client.user);
+    if (!permissions?.has(PermissionFlagsBits.ViewChannel)) return;
 
-        if (!permissions?.has(PermissionFlagsBits.EmbedLinks)) return interaction.reply({
-            content: client.messages.emojis["error"] + client.messages.noPermsEmbed,
-            ephemeral: true
-        });
-    }
+    if (!permissions?.has(PermissionFlagsBits.EmbedLinks)) return interaction.reply({
+        content: client.messages.emojis["error"] + client.messages.noPermsEmbed,
+        ephemeral: true
+    });
 
     if(interaction.isChatInputCommand()){
         const commandName = interaction.commandName;
